Make climber work with one side
